+++ make_messages.sh
@@ -1,0 +1,4 @@
+cd src/dash/
+django-admin.py makemessages -l hy
+django-admin.py makemessages -l nl
+django-admin.py makemessages -l ru