+++ src/dash/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'django-dash'
-__version__ = '0.2.4'
-__build__ = 0x000002
+__version__ = '0.3'
+__build__ = 0x000003
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'