+++ src/dash/contrib/layouts/bootstrap2/settings.py
@@ -1,0 +1,6 @@
+__all__ = ('IMAGE_CHOICES', 'IMAGE_CHOICES_WITH_EMPTY_OPTION')
+
+from dash.contrib.layouts.bootstrap2.conf import get_setting
+
+IMAGE_CHOICES = get_setting('IMAGE_CHOICES')
+IMAGE_CHOICES_WITH_EMPTY_OPTION = get_setting('IMAGE_CHOICES_WITH_EMPTY_OPTION')