+++ src/dash/contrib/plugins/image/helpers.py
@@ -1,8 +1,9 @@
-__all__ = ('handle_uploaded_file', 'get_crop_filter', 'delete_file')
+__all__ = ('handle_uploaded_file', 'get_crop_filter', 'delete_file', 'ensure_unique_filename')
 
 import os
 import glob
 import logging
+import uuid
 
 from django.conf import settings
 from django.core.files.base import File
@@ -15,16 +16,32 @@ logger = logging.getLogger(__file__)
 
 IMAGES_UPLOAD_DIR_ABSOLUTE_PATH = os.path.join(settings.MEDIA_ROOT, IMAGES_UPLOAD_DIR)
 
+def ensure_unique_filename(destination):
+    """
+    Makes sure filenames are never overwritten.
+
+    :param string destination:
+    :return string:
+    """
+    if os.path.exists(destination):
+        filename, extension = os.path.splitext(destination)
+        return "{0}_{1}{2}".format(filename, uuid.uuid4(), extension)
+    else:
+        return destination
+
 def handle_uploaded_file(image_file):
     """
     :param django.core.files.uploadedfile.InMemoryUploadedFile image_file:
     :return string: Path to the image (relative).
     """
     if isinstance(image_file, File):
-        with open(os.path.join(IMAGES_UPLOAD_DIR_ABSOLUTE_PATH, image_file.name), 'wb+') as destination:
+        destination_path = ensure_unique_filename(os.path.join(IMAGES_UPLOAD_DIR_ABSOLUTE_PATH, image_file.name))
+        image_filename = image_file.name
+        with open(destination_path, 'wb+') as destination:
+            image_filename = os.path.basename(destination.name)
             for chunk in image_file.chunks():
                 destination.write(chunk)
-        return os.path.join(IMAGES_UPLOAD_DIR, image_file.name)
+        return os.path.join(IMAGES_UPLOAD_DIR, image_filename)
     return image_file
 
 def get_crop_filter(fit_method):